--- src/main/java/BlackjackApplication.java
@@ -1,8 +1,0 @@
-import domain.card.Deck;
-
-public class BlackjackApplication {
-    public static void main(String[] args) {
-        BlackjackGame blackjackGame = new BlackjackGame(new Deck());
-        blackjackGame.run();
-    }
-}
