--- src/main/java/domain/GameStatistics.java
@@ -1,41 +1,0 @@
-package domain;
-
-import domain.participant.Participant;
-import java.util.EnumMap;
-import java.util.LinkedHashMap;
-import java.util.Map;
-
-public class GameStatistics {
-
-    private final Map<Participant, GameResult> playerResult;
-    private final Map<GameResult, Integer> dealerResult;
-
-    public GameStatistics() {
-        this.playerResult = new LinkedHashMap<>();
-        this.dealerResult = initialize();
-    }
-
-    private EnumMap<GameResult, Integer> initialize() {
-        EnumMap<GameResult, Integer> dealerResult = new EnumMap<>(GameResult.class);
-        for (GameResult gameResult : GameResult.values()) {
-            dealerResult.put(gameResult, 0);
-        }
-        return dealerResult;
-    }
-
-    public void addPlayerResult(Participant participant, GameResult gameResult) {
-        playerResult.put(participant, gameResult);
-    }
-
-    public void addDealerResult(GameResult gameResult) {
-        dealerResult.put(gameResult, dealerResult.getOrDefault(gameResult, 0) + 1);
-    }
-
-    public Map<Participant, GameResult> getPlayerResult() {
-        return playerResult;
-    }
-
-    public Map<GameResult, Integer> getDealerResult() {
-        return dealerResult;
-    }
-}
