--- src/main/java/util/Parser.java
@@ -1,16 +1,0 @@
-package util;
-
-import java.util.Arrays;
-import java.util.List;
-
-public class Parser {
-
-    private Parser() {
-    }
-
-    public static List<String> parseByDelimiter(String delimiter, String input) {
-        return Arrays.stream(input.split(delimiter))
-                .map(String::strip)
-                .toList();
-    }
-}
